Test library connection

+++ app/src/main/java/org/pesmypetcare/mypetcare/features/pets/Pet.java
@@ -21,9 +21,9 @@ public class Pet {
     private GenderType gender;
     private String breed;
     private String birthDate;
-    private float weight;
+    private double weight;
     private String pathologies;
-    private float recommendedDailyKiloCalories;
+    private double recommendedDailyKiloCalories;
     private int washFrequency;
     private User owner;
     private String previousName;
@@ -43,7 +43,7 @@ public class Pet {
 
         if (isMale(petInfo)) {
             this.gender = GenderType.Male;
-        } else if (isFemale(petInfo)){
+        } else if (isFemale(petInfo)) {
             this.gender = GenderType.Female;
         } else {
             this.gender = GenderType.Other;
@@ -147,7 +147,7 @@ public class Pet {
      * Get the weight of the pet.
      * @return The weight of the pet
      */
-    public float getWeight() {
+    public double getWeight() {
         return weight;
     }
 
@@ -155,7 +155,7 @@ public class Pet {
      * Set the weight of the pet.
      * @param weight The weight to set
      */
-    public void setWeight(float weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
     }
 
@@ -179,7 +179,7 @@ public class Pet {
      * Get the recommended daily kilo calories of the pet.
      * @return The recommended daily kilo calories of the pet
      */
-    public float getRecommendedDailyKiloCalories() {
+    public double getRecommendedDailyKiloCalories() {
         return recommendedDailyKiloCalories;
     }
 
@@ -187,7 +187,7 @@ public class Pet {
      * Set the recommended daily kilo calories of the pet.
      * @param recommendedDailyKiloCalories The recommended daily kilo calories of the pet to set
      */
-    public void setRecommendedDailyKiloCalories(float recommendedDailyKiloCalories) {
+    public void setRecommendedDailyKiloCalories(double recommendedDailyKiloCalories) {
         this.recommendedDailyKiloCalories = recommendedDailyKiloCalories;
     }
